refactor: split overgrown modules into single-responsibility files (#55, #58, #49+#54)
Task 1 (#55): Extract kwargs builder functions from editor_bridge.py into
editor_bridge_builders.py. All call sites import directly from the new
module; no re-export.

Task 2 (#58): Split symbol_tree.py (GUID index/build logic) into
symbol_tree_builder.py (258 lines). Extract cache layer from session.py
into session_cache.py (SessionCacheManager). session.py and symbol_tree.py
now ≤300 lines each.

Task 3 (#49+#54): Split mcp_server.py (2850→156 lines) into 11 focused
tool modules (mcp_tools_*.py) and mcp_helpers.py. mcp_server.py retained
as thin router/entrypoint calling register_*_tools(). All 70 MCP tools
preserved; no re-exports.

Tests: 1544 passed, 4 skipped. Ruff: all checks passed.
