+++ prefab_sentinel/editor_bridge.py
@@ -24,10 +24,6 @@ from prefab_sentinel.bridge_constants import (
     BRIDGE_WATCH_DIR_ENV,
     UNITY_TIMEOUT_SEC_ENV as BRIDGE_TIMEOUT_ENV,
 )
-from prefab_sentinel.editor_bridge_builders import (  # noqa: F401
-    build_create_empty_kwargs,
-    build_set_camera_kwargs,
-)
 from prefab_sentinel.json_io import dump_json, load_json
 
 PROTOCOL_VERSION = 1
