+++ prefab_sentinel/editor_bridge_builders.py
@@ -48,9 +48,9 @@ def build_create_empty_kwargs(
     name: str,
     parent_path: str = "",
     position: str = "",
-) -> dict[str, str]:
+) -> dict[str, Any]:
     """Build send_action kwargs from editor_create_empty parameters."""
-    kwargs: dict[str, str] = {"new_name": name}
+    kwargs: dict[str, Any] = {"new_name": name}
     if parent_path:
         kwargs["hierarchy_path"] = parent_path
     if position:
