+++ tests/test_editor_bridge.py
@@ -244,7 +244,7 @@ class TestSetCameraParams(unittest.TestCase):
     """Validate editor_set_camera parameter conversion."""
 
     def test_pivot_orbit_kwargs(self) -> None:
-        from prefab_sentinel.editor_bridge import build_set_camera_kwargs as _build_set_camera_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_set_camera_kwargs as _build_set_camera_kwargs
 
         kwargs = _build_set_camera_kwargs(
             pivot='{"x":0,"y":1.3,"z":0}',
@@ -260,7 +260,7 @@ class TestSetCameraParams(unittest.TestCase):
         self.assertNotIn("camera_look_at", kwargs)
 
     def test_position_look_at_kwargs(self) -> None:
-        from prefab_sentinel.editor_bridge import build_set_camera_kwargs as _build_set_camera_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_set_camera_kwargs as _build_set_camera_kwargs
 
         kwargs = _build_set_camera_kwargs(
             position='{"x":0,"y":1.5,"z":-1}',
@@ -271,7 +271,7 @@ class TestSetCameraParams(unittest.TestCase):
         self.assertNotIn("camera_pivot", kwargs)
 
     def test_position_yaw_pitch_kwargs(self) -> None:
-        from prefab_sentinel.editor_bridge import build_set_camera_kwargs as _build_set_camera_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_set_camera_kwargs as _build_set_camera_kwargs
 
         kwargs = _build_set_camera_kwargs(
             position='{"x":0,"y":1.5,"z":-1}',
@@ -286,13 +286,13 @@ class TestSetCameraParams(unittest.TestCase):
         self.assertNotIn("camera_look_at", kwargs)
 
     def test_omitted_params_excluded(self) -> None:
-        from prefab_sentinel.editor_bridge import build_set_camera_kwargs as _build_set_camera_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_set_camera_kwargs as _build_set_camera_kwargs
 
         kwargs = _build_set_camera_kwargs(yaw=180.0)
         self.assertEqual(kwargs, {"yaw": 180.0})
 
     def test_orthographic_passed(self) -> None:
-        from prefab_sentinel.editor_bridge import build_set_camera_kwargs as _build_set_camera_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_set_camera_kwargs as _build_set_camera_kwargs
 
         kwargs = _build_set_camera_kwargs(orthographic=1)
         self.assertEqual(kwargs["camera_orthographic"], 1)
@@ -302,7 +302,7 @@ class TestCreateEmptyKwargs(unittest.TestCase):
     """I4: build_create_empty_kwargs omits empty optional fields."""
 
     def test_name_only(self) -> None:
-        from prefab_sentinel.editor_bridge import build_create_empty_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_create_empty_kwargs
 
         result = build_create_empty_kwargs(name="Obj")
         self.assertEqual(result, {"new_name": "Obj"})
@@ -310,25 +310,25 @@ class TestCreateEmptyKwargs(unittest.TestCase):
         self.assertNotIn("property_value", result)
 
     def test_with_parent(self) -> None:
-        from prefab_sentinel.editor_bridge import build_create_empty_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_create_empty_kwargs
 
         result = build_create_empty_kwargs(name="Obj", parent_path="/Root")
         self.assertEqual(result, {"new_name": "Obj", "hierarchy_path": "/Root"})
 
     def test_with_position(self) -> None:
-        from prefab_sentinel.editor_bridge import build_create_empty_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_create_empty_kwargs
 
         result = build_create_empty_kwargs(name="Obj", position="1,2,3")
         self.assertEqual(result, {"new_name": "Obj", "property_value": "1,2,3"})
 
     def test_all_specified(self) -> None:
-        from prefab_sentinel.editor_bridge import build_create_empty_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_create_empty_kwargs
 
         result = build_create_empty_kwargs(name="Obj", parent_path="/Root", position="1,2,3")
         self.assertEqual(result, {"new_name": "Obj", "hierarchy_path": "/Root", "property_value": "1,2,3"})
 
     def test_empty_strings_omitted(self) -> None:
-        from prefab_sentinel.editor_bridge import build_create_empty_kwargs
+        from prefab_sentinel.editor_bridge_builders import build_create_empty_kwargs
 
         result = build_create_empty_kwargs(name="Obj", parent_path="", position="")
         self.assertEqual(result, {"new_name": "Obj"})
